day 8 user can buy send email with confirm password
(update 2)
send e-mail in separate thread

(update 3)
basket in separate model
added "/admin" to all admin fuctional
