+++ src/main/java/com/controller/DeleteProductServlet.java
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet("/delete/product")
+@WebServlet("/admin/delete/product")
 public class DeleteProductServlet extends HttpServlet {
 
     private static final ProductService productService = ProductServiceFactory.getInstance();
