+++ src/main/java/com/model/User.java
@@ -2,7 +2,6 @@ package com.model;
 
 import com.utils.IdGeneratorUtil;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class User {
@@ -83,26 +82,5 @@ public class User {
                 ", login='" + login + '\'' +
                 ", password='" + password + '\'' +
                 '}';
-    }
-
-    static class Basket {
-
-        private List<Product> productList;
-
-        public Basket() {
-            this.productList = new ArrayList<>();
-        }
-
-        public List<Product> getProductList() {
-            return productList;
-        }
-
-        public int getSize() {
-            return productList.size();
-        }
-
-        public void add(Product product) {
-            productList.add(product);
-        }
     }
 }
