+++ src/main/webapp/change_product.jsp
@@ -8,7 +8,7 @@
 
 ${error}
 
-<form action="/change/product?id=${productId}" method="post">
+<form action="/admin/change/product?id=${productId}" method="post">
     Название <input type="text" name="name" value="${oldName}"> <br>
     Описание <input type="text" name="description" value="${oldDescription}"> <br>
     Цена <input type="number" step="0.01" min="0" placeholder="0,00" name="price" value="${oldPrice}">
