+++ src/main/webapp/users.jsp
@@ -9,7 +9,7 @@
 
 <div align="left">
 
-    <button><a href="/add/user">Добавить пользователя</a></button>
+    <button><a href="/admin/add/user">Добавить пользователя</a></button>
 
     <table border="1">
         <tr>
@@ -23,10 +23,10 @@
                 <td>${user.login}</td>
                 <td>${user.password}</td>
                 <td>
-                    <button><a href="/change/user?id=${user.id}">Изменить</a></button>
+                    <button><a href="/admin/change/user?id=${user.id}">Изменить</a></button>
                 </td>
                 <td>
-                    <button><a href="/delete/user?id=${user.id}">Удалить</a></button>
+                    <button><a href="/admin/delete/user?id=${user.id}">Удалить</a></button>
                 </td>
             </tr>
         </c:forEach>
